refactor: 모듈 구조를 변경한다
- reducer와 store를 폴더 내로 이동한다
- 컴포넌트를 생성하는 함수를 사용하여 생성된 컴포넌트를 export한다
- entry point인 index.js 에서 각 컴포넌트를 동작시킨다
