+++ src/js/components/index.js
@@ -1,2 +1,4 @@
-import "./cashContainer.js";
-import "./lottoDetailContainer.js";
+import CashContainer from "./cashContainer.js";
+import LottoDetailContainer from "./lottoDetailContainer.js";
+
+export { CashContainer, LottoDetailContainer };
