+++ src/js/index.js
@@ -1,1 +1,4 @@
-import "./components/index.js";
+import { CashContainer, LottoDetailContainer } from "./components/index.js";
+
+CashContainer.init();
+LottoDetailContainer.init();
